Handle controller's resetting and connecting state
When a controller is in a resetting and connecting state, the device list becomes incomplete and broken.
The resulting error causes the engine validation logic to run into an error and leads to volume detachment.
To resolve this, if "connecting state" is returned, retry the validation logic in SPDK engine.

Longhorn 5744

Signed-off-by: Derek Su <derek.su@suse.com>